Refactored most of the REST API HTTP request validation code from the controllers to the Service abstract base class.  Added 3 custom exceptions for ServiceResponse<T> (BadRequestException, InternalErrorException, NotFoundException).  This yields better code reuse.  Controllers are a little bit leaner now.

+++ src/LiteBulb.OatShop.Shared/Services/Data/Service.cs
@@ -1,4 +1,5 @@
-﻿using LiteBulb.OatShop.Shared.Repositories;
+﻿using LiteBulb.OatShop.Shared.Exceptions;
+using LiteBulb.OatShop.Shared.Repositories;
 using Microsoft.Extensions.Logging;
 
 namespace LiteBulb.OatShop.Shared.Services.Data;
@@ -15,6 +16,8 @@ public abstract class Service<TModel, TId> : IService<TModel, TId>
         _repository = repository ?? throw new ArgumentNullException(nameof(repository));
     }
 
+    private bool IsDefaultValue(TId value) => EqualityComparer<TId>.Default.Equals(value, default);
+
     public virtual async Task<ServiceResponse<IReadOnlyList<TModel>>> GetAsync()
     {
         var result = await _repository.GetAsync();
@@ -22,7 +25,15 @@ public abstract class Service<TModel, TId> : IService<TModel, TId>
         if (result is null)
         {
             return new ServiceResponse<IReadOnlyList<TModel>>(true,
-                $"Error occurred while retrieving list of {_modelName} objects.  Result was null for some reason.");
+                $"Error occurred while retrieving list of {_modelName} objects.  Result was null for some reason.",
+                new InternalErrorException());
+        }
+
+        if (result.Count == 0)
+        {
+            return new ServiceResponse<IReadOnlyList<TModel>>(true,
+                "Empty list.",
+                new NotFoundException());
         }
 
         return new ServiceResponse<IReadOnlyList<TModel>>(result);
@@ -30,12 +41,27 @@ public abstract class Service<TModel, TId> : IService<TModel, TId>
 
     public virtual async Task<ServiceResponse<TModel>> GetAsync(TId id)
     {
+        if (id is null)
+        {
+            return new ServiceResponse<TModel>(true,
+                $"Id parameter cannot be null for Find By Id.",
+                new BadRequestException());
+        }
+
+        if (IsDefaultValue(id))
+        {
+            return new ServiceResponse<TModel>(true,
+                $"Id parameter cannot contain default value: '{id}' for Find By Id.",
+                new BadRequestException());
+        }
+
         var result = await _repository.GetAsync(id);
 
         if (result is null)
         {
             return new ServiceResponse<TModel>(true,
-                $"Error occurred while retrieving {_modelName} object with id '{id}'.  {_modelName} object was not found in the database.");
+                $"Error occurred while retrieving {_modelName} object with id '{id}'.  {_modelName} object was not found in the database.",
+                new NotFoundException());
         }
 
         return new ServiceResponse<TModel>(result);
@@ -43,12 +69,20 @@ public abstract class Service<TModel, TId> : IService<TModel, TId>
 
     public virtual async Task<ServiceResponse<TModel>> AddAsync(TModel model)
     {
+        if (model is null)
+        {
+            return new ServiceResponse<TModel>(true,
+                $"{_modelName} object payload cannot be null for Create.",
+                new BadRequestException());
+        }
+
         var result = await _repository.AddAsync(model);
 
         if (result is null)
         {
             return new ServiceResponse<TModel>(true,
-                $"Error occurred while adding a {_modelName} object to database.  Result returned by add process was null for some reason.");
+                $"Error occurred while adding a {_modelName} object to database.  Result returned by add process was null for some reason.",
+                new InternalErrorException());
         }
 
         return new ServiceResponse<TModel>(result);
@@ -56,18 +90,41 @@ public abstract class Service<TModel, TId> : IService<TModel, TId>
 
     public virtual async Task<ServiceResponse<int>> UpdateAsync(TId id, TModel model)
     {
+        if (id is null)
+        {
+            return new ServiceResponse<int>(true,
+                $"Id parameter cannot be null for Update.",
+                new BadRequestException());
+        }
+
+        if (IsDefaultValue(id))
+        {
+            return new ServiceResponse<int>(true,
+                $"Id parameter cannot contain default value: '{id}' for Update.",
+                new BadRequestException());
+        }
+
+        if (model is null)
+        {
+            return new ServiceResponse<int>(true,
+                $"{_modelName} object payload cannot be null for Update.",
+                new BadRequestException());
+        }
+
         var affectedCount = await _repository.UpdateAsync(id, model);
 
         if (affectedCount is null || !affectedCount.HasValue)
         {
             return new ServiceResponse<int>(true,
-               $"Error occurred while updating a {_modelName} object in the database.  {_modelName} object with id '{id}' was not found in the database.");
+               $"Error occurred while updating a {_modelName} object in the database.  {_modelName} object with id '{id}' was not found in the database.",
+               new NotFoundException());
         }
 
         if (affectedCount < 1)
         {
             return new ServiceResponse<int>(true,
-               $"Error occurred while updating a {_modelName} object in the database.  The affected record count is '{affectedCount}' which is invalid.");
+               $"Error occurred while updating a {_modelName} object in the database.  The affected record count is '{affectedCount}' which is invalid.",
+               new InternalErrorException());
         }
 
         return new ServiceResponse<int>(affectedCount.Value);
@@ -75,18 +132,34 @@ public abstract class Service<TModel, TId> : IService<TModel, TId>
 
     public virtual async Task<ServiceResponse<int>> DeleteAsync(TId id)
     {
+        if (id is null)
+        {
+            return new ServiceResponse<int>(true,
+                $"Id parameter cannot be null for Delete By Id.",
+                new BadRequestException());
+        }
+
+        if (IsDefaultValue(id))
+        {
+            return new ServiceResponse<int>(true,
+                $"Id parameter cannot contain default value: '{id}' for Delete By Id.",
+                new BadRequestException());
+        }
+
         var affectedCount = await _repository.DeleteAsync(id);
 
         if (affectedCount is null || !affectedCount.HasValue)
         {
             return new ServiceResponse<int>(true,
-               $"Error occurred while deleting a {_modelName} object from the database.  {_modelName} object with id '{id}' was not found in the database.");
+               $"Error occurred while deleting a {_modelName} object from the database.  {_modelName} object with id '{id}' was not found in the database.",
+               new NotFoundException());
         }
 
         if (affectedCount < 1)
         {
             return new ServiceResponse<int>(true,
-               $"Error occurred while deleting a {_modelName} object from the database.  The affected record count is '{affectedCount}' which is invalid.");
+               $"Error occurred while deleting a {_modelName} object from the database.  The affected record count is '{affectedCount}' which is invalid.",
+               new InternalErrorException());
         }
 
         return new ServiceResponse<int>(affectedCount.Value);
